Unify utf-8 and text into text
