+++ tests/expected/decode/test.utf8.stdout
@@ -1,3 +1,4 @@
 ├── data <- _ |...| _ :=
-│   └── utf8 <- utf8.string := "この🦀は擂り身ではなく、本物のカニです。\n"
+│   └── text <- text.string :=
+│       └── utf8 <- text.string.utf8 := "この🦀は擂り身ではなく、本物のカニです。\n"
 └── end <- end-of-input
